# Implement caching and batching for performance