--- a/packages/ilmomasiina-backend/src/routes/admin/events/createEvent.ts
+++ b/packages/ilmomasiina-backend/src/routes/admin/events/createEvent.ts
@@ -7,7 +7,8 @@ import { getSequelize } from "../../../models";
 import { Event } from "../../../models/event";
 import { Question } from "../../../models/question";
 import { Quota } from "../../../models/quota";
-import { eventDetailsForAdmin } from "../../events/getEventDetails";
+import { basicEventInfoCached, eventDetailsForAdmin, eventDetailsForUserCached } from "../../events/getEventDetails";
+import { eventsListForUserCached } from "../../events/getEventsList";
 import { toDate } from "../../utils";
 
 export default async function createEvent(
@@ -56,6 +57,10 @@ export default async function createEvent(
 
     return created;
   });
+
+  eventsListForUserCached.invalidate();
+  basicEventInfoCached.invalidate();
+  eventDetailsForUserCached.invalidate();
 
   const eventDetails = await eventDetailsForAdmin(event.id);