--- a/packages/ilmomasiina-backend/src/routes/events/getEventsList.ts
+++ b/packages/ilmomasiina-backend/src/routes/events/getEventsList.ts
@@ -7,6 +7,7 @@ import { Event } from "../../models/event";
 import { Quota } from "../../models/quota";
 import { Signup } from "../../models/signup";
 import { ascNullsFirst } from "../../models/util";
+import createCache from "../../util/cache";
 import { InitialSetupNeeded, isInitialSetupDone } from "../admin/users/createInitialUser";
 import { StringifyApi } from "../utils";
 
@@ -20,6 +21,43 @@ function eventOrder(): Order {
   ];
 }
 
+export const eventsListForUserCached = createCache({
+  maxAgeMs: 1000,
+  maxPendingAgeMs: 2000,
+  async get(category?: string) {
+    const where = category ? { category } : {};
+
+    const events = await Event.scope("user").findAll({
+      attributes: eventListEventAttrs,
+      where: { listed: true, ...where },
+      // Include quotas of event and count of signups
+      include: [
+        {
+          model: Quota,
+          attributes: ["id", "title", "size", [fn("COUNT", col("quotas->signups.id")), "signupCount"]],
+          include: [
+            {
+              model: Signup.scope("active"),
+              required: false,
+              attributes: [],
+            },
+          ],
+        },
+      ],
+      group: [col("event.id"), col("quotas.id")],
+      order: eventOrder(),
+    });
+
+    return events.map((event) => ({
+      ...event.get({ plain: true }),
+      quotas: event.quotas!.map((quota) => ({
+        ...quota.get({ plain: true }),
+        signupCount: Number(quota.signupCount),
+      })),
+    }));
+  },
+});
+
 export async function getEventsListForUser(
   this: FastifyInstance<any, any, any, any, any>,
   request: FastifyRequest<{ Querystring: EventListQuery }>,
@@ -30,38 +68,11 @@ export async function getEventsListForUser(
     throw new InitialSetupNeeded("Initial setup of Ilmomasiina is needed.");
   }
 
-  const events = await Event.scope("user").findAll({
-    attributes: eventListEventAttrs,
-    where: { listed: true, ...request.query },
-    // Include quotas of event and count of signups
-    include: [
-      {
-        model: Quota,
-        attributes: ["id", "title", "size", [fn("COUNT", col("quotas->signups.id")), "signupCount"]],
-        include: [
-          {
-            model: Signup.scope("active"),
-            required: false,
-            attributes: [],
-          },
-        ],
-      },
-    ],
-    group: [col("event.id"), col("quotas.id")],
-    order: eventOrder(),
-  });
-
-  const res = events.map((event) => ({
-    ...event.get({ plain: true }),
-    quotas: event.quotas!.map((quota) => ({
-      ...quota.get({ plain: true }),
-      signupCount: Number(quota.signupCount),
-    })),
-  }));
-
+  const res = await eventsListForUserCached(request.query.category);
   reply.status(200);
   return res as StringifyApi<typeof res>;
 }
+
 export async function getEventsListForAdmin(
   request: FastifyRequest<{ Querystring: EventListQuery }>,
   reply: FastifyReply,