--- a/packages/ilmomasiina-backend/src/routes/signups/computeSignupPosition.ts
+++ b/packages/ilmomasiina-backend/src/routes/signups/computeSignupPosition.ts
@@ -28,21 +28,17 @@ async function sendPromotedFromQueueMail(signup: Signup, eventId: Event["id"]) {
   await EmailService.sendPromotedFromQueueMail(signup.email, signup.language, params);
 }
 
-/**
- * Updates the status and position attributes on all signups in the given event. Also sends "promoted from queue"
- * emails to affected users. Returns the new statuses for all signups.
- *
- * By default, recomputations can move signups into the queue. This ensures that we don't cause random errors for
- * ordinary users. `moveSignupsToQueue = false` is passed if a warning can be shown (i.e. in admin-side editors).
- */
-export async function refreshSignupPositions(
+/** Internal, non-batched version. See below for explanation of what this does. */
+async function refreshSignupPositionsInternal(
   eventRef: Event,
   transaction?: Transaction,
   moveSignupsToQueue: boolean = true,
 ): Promise<Signup[]> {
   // Wrap in transaction if not given
   if (!transaction) {
-    return getSequelize().transaction(async (trans) => refreshSignupPositions(eventRef, trans));
+    return getSequelize().transaction(async (trans) =>
+      refreshSignupPositionsInternal(eventRef, trans, moveSignupsToQueue),
+    );
   }
 
   // Lock event to prevent simultaneous changes
@@ -138,6 +134,68 @@ export async function refreshSignupPositions(
   );
 
   return result.map(({ signup }) => signup);
+}
+
+const refreshQueues = new Map<Event["id"], { ongoing?: Promise<unknown>; queued?: Promise<Signup[]> }>();
+
+/**
+ * Updates the status and position attributes on all signups in the given event. Also sends "promoted from queue"
+ * emails to affected users. Returns the new statuses for all signups.
+ *
+ * By default, recomputations can move signups into the queue. This ensures that we don't cause random errors for
+ * ordinary users. `moveSignupsToQueue = false` is passed if a warning can be shown (i.e. in admin-side editors).
+ *
+ * This action is batched due to database locking - if a call is already ongoing, the next call will be delayed and
+ * performed only once for all calls performed during a previous operation.
+ */
+export async function refreshSignupPositions(
+  eventRef: Event,
+  transaction?: Transaction,
+  moveSignupsToQueue: boolean = true,
+): Promise<Signup[]> {
+  // If a transaction is passed, we need to do the refresh within that transaction.
+  // It may need to wait for other transactions, but let's handle that on the DB level.
+  if (transaction) {
+    return refreshSignupPositionsInternal(eventRef, transaction, moveSignupsToQueue);
+  }
+
+  let queue = refreshQueues.get(eventRef.id);
+  if (!queue) {
+    queue = {};
+    refreshQueues.set(eventRef.id, queue);
+  }
+
+  // If a another refresh is already ongoing and another is queued, reuse the last queued one,
+  // since it will be executed after this moment anyway.
+  if (queue.queued) return queue.queued;
+
+  // Otherwise, if another refresh is already ongoing, queue a new one after it completes.
+  if (queue.ongoing) {
+    const queued = queue.ongoing
+      .catch(() => {
+        // ignore errors, we always want to run after the ongoing refresh finishes
+      })
+      .then(async () => {
+        queue!.ongoing = queued;
+        queue!.queued = undefined;
+        try {
+          return await refreshSignupPositionsInternal(eventRef, transaction, moveSignupsToQueue);
+        } finally {
+          if (queue!.ongoing === queued) queue!.ongoing = undefined;
+        }
+      });
+    queue.queued = queued;
+    return queued;
+  }
+
+  // Otherwise, immediately perform a refresh.
+  const promise = refreshSignupPositionsInternal(eventRef, transaction, moveSignupsToQueue);
+  queue.ongoing = promise;
+  try {
+    return await promise;
+  } finally {
+    if (queue.ongoing === promise) queue.ongoing = undefined;
+  }
 }
 
 /**