- Cache event list and event details DB fetches for users
- Batch recomputing signup statuses/positions

Ran `loadTest.js` on my laptop with ~100 events and ~10k signups generated by `fillDatabase.ts`.

Before:
```
     http_req_duration_events_list.........: avg=1.33s    min=53.44ms  med=1.33s    max=5.1s    p(90)=1.39s    p(95)=1.42s   
     http_req_duration_individual_events...: avg=730.41ms min=457.2ms  med=693.44ms max=2.95s   p(90)=769.92ms p(95)=806.47ms
     http_req_duration_sign_up_to_quota....: avg=2.37s    min=979.54ms med=2.21s    max=3.61s   p(90)=3.41s    p(95)=3.52s   
     http_reqs_events_list.................: 2293   24.494281/s
     http_reqs_individual_events...........: 4166   44.502038/s
     http_reqs_sign_up_to_quota............: 1338   14.292781/s
```

After:
```
     http_req_duration_events_list.........: avg=232.24ms min=5.09ms   med=214.34ms max=1.16s   p(90)=287.86ms p(95)=305.97ms
     http_req_duration_individual_events...: avg=42.53ms  min=868.91µs med=33.08ms  max=2.48s   p(90)=64.8ms   p(95)=101.16ms
     http_req_duration_sign_up_to_quota....: avg=529.02ms min=143.46ms med=489.39ms max=1.54s   p(90)=730.67ms p(95)=753.94ms
     http_reqs_events_list.................: 12948  142.9286/s
     http_reqs_individual_events...........: 70335  776.404317/s
     http_reqs_sign_up_to_quota............: 5757   63.549579/s
```